requirements.txt

plexapi
python-dotenv
